Implement non-flushable output buffers and fix RPC early flush
If EnableEarlyFlush is true, which is the default, then flushing the
output in an RPC request prior to request termination, explicitly or
implicitly, will cause the partial output to be immediately sent,
breaking the JSON serialization. The dynamic content cache is also
broken in the same way.

So, following PHP PHP, implement the PHP_OUTPUT_HANDLER_FLUSHABLE flag
to ob_start(). Move the obStart() in hphp_context_init() to the two
callers that actually need it, and clear the flushable flag in the RPC
case and the cacheable case.

Also, ext_std_string.h had an incorrect inclusion guard which broke
compilation of this patch, so I fixed it.