Fix protected output buffer flushing
Summary

1. Implement non-flushable output buffers and fix RPC early flush
2. Fix HEAD handling, don't require buffering
3. Migrate HTTP/1.0 handling to new flushable flag
4. Allow obFlush() of top-level output buffer

Details

1) Implement non-flushable output buffers and fix RPC early flush

If EnableEarlyFlush is true, which is the default, then flushing the
output in an RPC request prior to request termination, explicitly or
implicitly, will cause the partial output to be immediately sent,
breaking the JSON serialization. The dynamic content cache is also
broken in the same way.

So, following PHP PHP, implement the PHP_OUTPUT_HANDLER_FLUSHABLE flag
to ob_start(). Move the obStart() in hphp_context_init() to the two
callers that actually need it, and clear the flushable flag in the RPC
case and the cacheable case.

Also, ext_std_string.h had an incorrect inclusion guard which broke
compilation of this patch, so I fixed it.

2) Fix HEAD handling, don't require buffering

On a HEAD request, don't require the complete output to be buffered,
since support for large streaming output is a long-term goal.

The theory seems to have been that if you buffer HEAD requests, you can
send a Content-Length header. This could be more efficiently implemented
by just counting the bytes, but I took a different approach. I don't
think it is correct to send a Content-Length header when you wouldn't
send one for a GET request, so I just had it send headers as normal in
response to the first stdout write. If the chunked=false, we send a
Content-Length header for both GET and HEAD. This being done, the body
of the response to a HEAD request can be safely discarded.

In the Proxygen server, HEAD requests were totally broken: the body was
sent. So I had to implement HEAD handling from scratch.

In the FastCGI server, it is the webserver's responsibility to do HEAD
handling, and the protocol gives no way to indicate content length
without just sending the content. So buffering was never necessary and
no change is needed.

Tested Proxygen chunked and unchunked. Did not test the deprecated
libevent server since I assume it will soon be deleted in favour of
Proxygen.

3) Migrate HTTP/1.0 handling to new flushable flag

Migrate HTTP/1.0 handling from a special case in
ExecutionContext::flush() to the new k_PHP_OUTPUT_HANDLER_FLUSHABLE
flag, set by the obStart() caller in HttpRequestHandler, like dynamic
content caching and RPC serialization. ExecutionContext is not the right
place for HTTP protocol handling, and doing this clears the way for
future changes to flushing policy in service of streaming output.

4) Allow obFlush() of top-level output buffer

The dependencies on an infinite size, unflushable top-level output buffer
have been removed or migrated to use k_PHP_OUTPUT_HANDLER_FLUSHABLE. So,
allow the top-level output buffer to be flushed with obFlush().

* Add a new helper function ExecutionContext::writeTransport() which
  sends data to the transport if there is one, or to stdout otherwise.
  Otherwise the logic would be duplicated in a few places.
* Do not conditionally call fflush() after write(), that is pointless
  since fflush() only flushes the userspace output buffer and write()
  does not use the userspace output buffer. Instead, call fflush()
  unconditionally before write(), so that if any external libraries have
  used the FILE* functions previously, the end result will be in order.
* Since obFlush() no longer fails when it is called on the last output
  buffer, it is necessary to use the return value from obEnd() instead
  to terminate the obFlushAll() loop.