Fix HEAD handling, don't require buffering
On a HEAD request, don't require the complete output to be buffered,
since support for large streaming output is a long-term goal.

The theory seems to have been that if you buffer HEAD requests, you can
send a Content-Length header. This could be more efficiently implemented
by just counting the bytes, but I took a different approach. I don't
think it is correct to send a Content-Length header when you wouldn't
send one for a GET request, so I just had it send headers as normal in
response to the first stdout write. If the chunked=false, we send a
Content-Length header for both GET and HEAD. This being done, the body
of the response to a HEAD request can be safely discarded.

In the Proxygen server, HEAD requests were totally broken: the body was
sent. So I had to implement HEAD handling from scratch.

In the FastCGI server, it is the webserver's responsibility to do HEAD
handling, and the protocol gives no way to indicate content length
without just sending the content. So buffering was never necessary and
no change is needed.

Tested Proxygen chunked and unchunked. Did not test the deprecated
libevent server since I assume it will soon be deleted in favour of
Proxygen.